// components/layout/src/gqlTypes.d.ts
export type JSON = any;
// ...
export enum EmployeeType {
  AD = 'AD',
  RE = 'RE',
  HC = 'HC',
  IN = 'IN',
  SH = 'SH',
  TE = 'TE',
}
// ...
  export type Variables = {};

  export type Query = {
    __typename?: 'Query';
    dashboard: Dashboard;
  };

  export type Dashboard = {
    __typename?: 'Dashboard';
    id?: string | null;
    isSpoofing?: boolean | null;
    isTrialCompany?: boolean | null;
    switches: JSON;
    permission?: JSON | null;
    employee?: Employee | null;
    zAppInstallSubscriptions?: (ZAppInstallSubscriptions | null)[] | null;
  };

  export type Employee = {
    __typename?: 'AllEmployee';
    id?: string | null;
    type?: EmployeeType | null;
  };

  export type ZAppInstallSubscriptions = {
    __typename?: 'ZAppInstallSubscription';
    status?: number | null;
    inheritedStatus?: number | null;
    preferences?: JSON | null;
    appInstall?: AppInstall | null;
  };
// ...
export namespace TopNavBarQuery {
  export type Variables = {};

  export type Query = {
    __typename?: 'Query';
    dashboard: Dashboard;
  };

  export type Dashboard = {
    __typename?: 'Dashboard';
    id?: string | null;
    isMTAUser?: boolean | null;
    isMTAPartnerUser?: boolean | null;
    isConsoleUser?: boolean | null;
    isSpoofing?: boolean | null;
    isDemoAccount?: boolean | null;
    isTrialCompany?: boolean | null;
    isTrialPasswordless?: boolean | null;
    isTrialSales?: boolean | null;
    switches: JSON;
    permission?: JSON | null;
    canSeeDemoCenter?: boolean | null;
    user?: User | null;
    employee?: Employee | null;
    partner?: Partner | null;
    company?: Company | null;
  };

  export type User = {
    __typename?: 'User';
    id?: string | null;
    first_name?: string | null;
    last_name?: string | null;
  };

  export type Employee = {
    __typename?: 'AllEmployee';
    id?: string | null;
    first_name?: string | null;
    last_name?: string | null;
    photoUrl?: string | null;
  };

  export type Partner = {
    __typename?: 'Partner';
    id?: string | null;
    companyId?: string | null;
  };

  export type Company = {
    __typename?: 'Company';
    id?: string | null;
    regEmployeeCount?: number | null;
  };
}